feat/#59 이미지 뷰 생성
