--- PodongPodong/PodongPodong/Presentations/ImageViews/AlarmSettingView.swift
@@ -1,21 +1,0 @@
-//
-//  AlarmSettingView.swift
-//  PodongPodong
-//
-//  Created by 김서현 on 6/10/25.
-//
-
-// MARK: - 마이페이지 > 알림 > 알림 설정 View
-import SwiftUI
-
-struct AlarmSettingView: View {
-    var body: some View {
-        Image("AlarmSettingViewImage")
-            .resizable()
-            .scaledToFit()
-    }
-}
-
-#Preview {
-    AlarmSettingView()
-}
