--- PodongPodong/PodongPodong/Presentations/ImageViews/EditProfileView.swift
@@ -1,22 +1,0 @@
-//
-//  EditProfileView.swift
-//  PodongPodong
-//
-//  Created by 김서현 on 6/10/25.
-//
-
-// MARK: - 마이페이지 > 프로필 수정 버튼 View
-
-import SwiftUI
-
-struct EditProfileView: View {
-    var body: some View {
-        Image("EditProfileViewImage")
-            .resizable()
-            .scaledToFit()
-    }
-}
-
-#Preview {
-    EditProfileView()
-}
