--- PodongPodong/PodongPodong/Presentations/ImageViews/InterestedPartiesView.swift
@@ -1,20 +1,0 @@
-//
-//  InterestedPartiesView.swift
-//  PodongPodong
-//
-//  Created by 김서현 on 6/9/25.
-//
-// MARK: - 마이페이지 > 관심 파티 목록 View
-import SwiftUI
-
-struct InterestedPartiesView: View {
-    var body: some View {
-        Image("InterestedPartiesViewImage")
-            .resizable()
-            .scaledToFit()
-    }
-}
-
-#Preview {
-    InterestedPartiesView()
-}
