--- PodongPodong/PodongPodong/Presentations/ImageViews/KeywordRegistrationView.swift
@@ -1,20 +1,0 @@
-//
-//  KeywordRegistrationView.swift
-//  PodongPodong
-//
-//  Created by 김서현 on 6/10/25.
-//
-// MARK: - 마이페이지 > 알림 > 키워드 알림 View
-import SwiftUI
-
-struct KeywordRegistrationView: View {
-    var body: some View {
-        Image("KeywordRegistrationViewImage")
-            .resizable()
-            .scaledToFit()
-    }
-}
-
-#Preview {
-    KeywordRegistrationView()
-}
