--- PodongPodong/PodongPodong/Presentations/ImageViews/ParticipatedView.swift
@@ -1,21 +1,0 @@
-//
-//  ParticipatedView.swift
-//  PodongPodong
-//
-//  Created by 김서현 on 6/9/25.
-//
-
-// MARK: - 마이페이지> 참여한 파티 View
-import SwiftUI
-
-struct ParticipatedView: View {
-    var body: some View {
-        Image("ParticipatedViewImage")
-            .resizable()
-            .scaledToFit()
-    }
-}
-
-#Preview {
-    ParticipatedView()
-}
