+++ PodongPodong/PodongPodong/Resource/Assets.xcassets/MyReviewImage.imageset/Contents.json
@@ -9,7 +9,7 @@
       "scale" : "2x"
     },
     {
-      "filename" : "KeywordRegistrationViewImage.png",
+      "filename" : "MyReviewImage.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
